LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-destination-getState-navigate-expected.txt


FAIL navigate event destination.getState() should be the state given to navigate() assert_equals: expected (object) null but got (undefined) undefined
